backend/requirements.txt
# FastAPI and ASGI server
# @url: https://fastapi.tiangolo.com/
fastapi==0.109.2
# @url: https://www.uvicorn.org/
uvicorn==0.27.1
# @url: https://github.com/andrew-d/python-multipart
python-multipart==0.0.9

# Database
# @url: https://www.sqlalchemy.org/
sqlalchemy==2.0.27
# @url: https://alembic.sqlalchemy.org/
alembic==1.13.1
# @url: https://www.psycopg.org/
psycopg2-binary==2.9.9

# Environment variables
# @url: https://python-dotenv.readthedocs.io/
python-dotenv==1.0.1

# Authentication
# @url: https://python-jose.readthedocs.io/
python-jose==3.3.0
# @url: https://passlib.readthedocs.io/
passlib[bcrypt]==1.7.4
# @url: https://python-bcrypt.readthedocs.io/
bcrypt==3.2.2

# Testing
# @url: https://docs.pytest.org/
pytest==8.0.1
# @url: https://coverage.readthedocs.io/
coverage==7.4.1

# Utilities
# @url: https://pydantic-docs.helpmanual.io/
pydantic[email]==2.6.1
# @url: https://pydantic-settings.helpmanual.io/
pydantic-settings==2.2.1
# @url: https://email-validator.readthedocs.io/
email-validator==2.2.0
# @url: https://requests.readthedocs.io/
requests==2.31.0
# @url: https://www.python-httpx.org/
httpx==0.26.0

# Google API
# @url: https://github.com/googleapis/google-auth-library-python
google-auth==2.28.1
# @url: https://github.com/googleapis/google-api-python-client
google-api-python-client==2.120.0

# RAG dependencies
# @url: https://numpy.org/
numpy==1.26.3
# @url: https://github.com/matthewwithanm/python-markdownify
markdownify==0.11.6
# @url: https://github.com/py-pdf/pypdf
PyPDF2==3.0.1
# @url: https://github.com/python-openxml/python-docx
python-docx==1.1.0